+++ src/bls12_381/fr.rs
@@ -23,7 +23,7 @@ const INV: u64 = 0xfffffffeffffffff;
 const GENERATOR: FrRepr = FrRepr([0xefffffff1, 0x17e363d300189c0f, 0xff9c57876f8457b0, 0x351332208fc5a8c4]);
 
 // 2^s * t = MODULUS - 1 with t odd
-const S: usize = 32;
+const S: u32 = 32;
 
 // 2^s root of unity computed by GENERATOR^t
 const ROOT_OF_UNITY: FrRepr = FrRepr([0xb9b58d8c5f0e466a, 0x5b1b4c801819d7ec, 0xaf53ae352a31e64, 0x5bf3adda19e9b27b]);
@@ -264,7 +264,7 @@ impl PrimeField for Fr {
         Fr(GENERATOR)
     }
 
-    fn s() -> usize {
+    fn s() -> u32 {
         S
     }
